+++ rollup.config.js
@@ -21,9 +21,9 @@ let plugins = [
       // relative to the current directory, or the name
       // of a module in node_modules
       'radium': [ 'Style', 'StyleRoot' ],
-      'react': [ 'Component', 'PropTypes', 'Children', 'createElement' ],
+      'react': [ 'PureComponent', 'Component', 'PropTypes', 'Children', 'createElement' ],
       'react-dom': [ 'unstable_renderSubtreeIntoContainer', 'unmountComponentAtNode' ],
-      'js-yaml': [ 'safeLoad', 'CORE_SCHEMA' ]
+      'js-yaml': [ 'safeLoad', 'CORE_SCHEMA', 'Type', 'Schema' ]
     }
   }),
   babel({
@@ -45,5 +45,8 @@ export default {
   globals: {
     'babel-standalone': 'Babel'
   },
+  external: [
+    'babel-standalone'
+  ],
   banner: '/*! Catalog ' + version + ' http://interactivethings.github.io/catalog/ */'
 };