+++ src/utils/parseSpecimenBody.js
@@ -1,6 +1,4 @@
-import {safeLoad, CORE_SCHEMA} from 'js-yaml';
-
-const yamlOptions = {schema: CORE_SCHEMA};
+import {safeLoad, CORE_SCHEMA, Type, Schema} from 'js-yaml';
 
 const defaultMapBodyToProps = (parsedBody, rawBody) => parsedBody || rawBody;
 
@@ -16,9 +14,24 @@ const splitText = (text) => {
     [void 0, text];
 };
 
-const parseYaml = (str) => {
+const parseYaml = (str, imports) => {
   let parsed;
   try {
+    const ImportType = new Type('!import', {
+      kind: 'scalar',
+      // TODO: Gracefully handle missing imports
+      // resolve(key) {
+      //   return imports.hasOwnProperty(key);
+      // },
+      construct(key) {
+        return imports[key];
+      }
+    });
+
+    const yamlOptions = {
+      schema: Schema.create(CORE_SCHEMA, [ImportType])
+    };
+
     parsed = safeLoad(str, yamlOptions);
   } catch (e) {
     parsed = void 0;
@@ -26,14 +39,14 @@ const parseYaml = (str) => {
   return typeof parsed === 'string' ? void 0 : parsed;
 };
 
-export const parseSpecimenYamlBody = (_mapBodyToProps) => (body = '') => {
+export const parseSpecimenYamlBody = (_mapBodyToProps) => (body = '', imports = {}) => {
   const mapBodyToProps = _mapBodyToProps || defaultMapBodyToProps;
-  return mapBodyToProps(parseYaml(body), body);
+  return mapBodyToProps(parseYaml(body, imports), body);
 };
 
-export const parseSpecimenBody = (_mapBodyToProps) => (body = '') => {
+export const parseSpecimenBody = (_mapBodyToProps) => (body = '', imports = {}) => {
   const mapBodyToProps = _mapBodyToProps || defaultMapBodyToProps;
   const splitBody = splitText(body);
   const [props, children] = splitBody;
-  return mapBodyToProps({...parseYaml(props), children}, body);
+  return mapBodyToProps({...parseYaml(props, imports), children}, body);
 };